pyproject.toml
[tool.poetry]
name = "src"
version = "0.1.0"
description = ""
authors = ["leeyeoreum02 <leeyeoreum01@gmail.com>"]
readme = "README.md"
# package-mode = false

[tool.poetry.dependencies]
python = "^3.8"
torch = {version = "2.2.1+cu121", source = "torch"}
torchvision = "^0.17.1"
numpy="^1"
scipy = "^1.6.2"
tqdm = "^4.66.4"
open-clip-torch = "2.0.2"
rich = "^13.7.1"
matplotlib = "3.7"
seaborn = "^0.13.2"
wandb = "^0.18.5"
datasets = "^3.0.2"
h5py = "3.10"
gdown = "^5.2.0"
scikit-learn = "^1.3.2"

[tool.poetry.group.dev.dependencies]
black = "^24.2.0"
pytest = "^8.1.1"
isort = "^5.13.2"
mypy = "^1.9.0"
ipython = "~8.12"
ipykernel = "^6.29.4"

[[tool.poetry.source]]
name = "PyPI"
priority = "primary"

[[tool.poetry.source]]
name = "torch"
url = "https://download.pytorch.org/whl/cu121"
priority = "explicit"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
addopts = "-s"
testpaths = ["tests"]

[tool.black]
line-length = 120
target-version = ["py38", "py39", "py310"]
exclude = """.venv|venv|.env"""

[tool.isort]
skip = [".gitignore"]

[tool.flake8]
exclude = ".venv"
max-line-length = 120

[tool.mypy]
exclude = ["tests"]
